% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{avg_slope_of_bspline}
\alias{avg_slope_of_bspline}
\title{Computes the average slope of a B-spline based on its coefficients.}
\usage{
avg_slope_of_bspline(knots, coef, l = 3)
}
\arguments{
\item{knots}{Vector of spline knots.}

\item{coef}{Vector of spline coefficients.}

\item{l}{Order of the spline. The default is \code{l = 3} for a cubic spline.}
}
\value{
Average slope of the spline, a scalar.
}
\description{
Assumes equidistant knots.
}
