% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{bg.go}
\alias{bg.go}
\title{Search GO Terms for Background Set}
\usage{
bg.go(ids)
}
\arguments{
\item{ids}{either a vector containing the UniProt IDs of the background set or the path to the txt file containing the list of IDs acting as background.}
}
\value{
Returns a dataframe with two columns (Uniprot ID, GO terms) and as many rows as different proteins there are in the input set.
}
\description{
Searches the GO terms of the protein contained in a given set.
}
\examples{
\dontrun{bg.go(c('P01009', 'P01374', 'Q86UP4'))}
}
\references{
Rhee et al. (2008) Nature Reviews Genetics 9:509–515.
}
\seealso{
search.go(), term.go(), get.go(), go.enrich(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
