% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmScan.R
\name{dis.scan}
\alias{dis.scan}
\title{Scan a Protein in Search of Disease-Related PTM Sites}
\usage{
dis.scan(up_id)
}
\arguments{
\item{up_id}{a character string corresponding to the UniProt ID.}
}
\value{
Returns a dataframe where each row corresponds to a residue, and the columns inform about the disease-related modifications.
}
\description{
Scans the indicated protein in search of disease-related PTM sites.
}
\examples{
\dontrun{dis.scan('P31749')}
}
\references{
Hornbeck et al. Nucleic Acids Res. 2019 47:D433-D441, (PMID: 30445427).
}
\seealso{
meto.scan(), ac.scan(), me.scan(), ub.scan(), su.scan(), gl.scan(), sni.scan(), ni.scan(), ptm.scan(), reg.scan(), p.scan()
}
\author{
Juan Carlos Aledo
}
