% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.R
\name{dtramo}
\alias{dtramo}
\alias{ptramo}
\title{The Transformation Distribution.}
\usage{
dtramo(z, z_deriv, reference_density = stats::dnorm, log = FALSE, ...)

ptramo(z, reference_cdf = stats::pnorm, ...)
}
\arguments{
\item{z, z_deriv}{Numeric arrays of transformed values and transformation
derivative evaluations.}

\item{reference_density, reference_cdf}{Density and cumulative distribution
function of the reference distribution.}

\item{log}{Logical; if \code{TRUE}, log densities are returned.}

\item{...}{Additional arguments passed on to \code{reference_density}}
}
\value{
Numeric array of density evaluations.
}
\description{
Density and cumulative distribution function of the transformation
distribution.
}
\examples{
knots <- equidistant_knots(c(-4, 4), p = 10)
approx_basis_and_deriv <- get_approx_basis_and_deriv_fun(knots, ngrid = 1000)
spline <- get_extrap_bspline_and_deriv(approx_basis_and_deriv)

# evaluate spline
x <- rnorm(n = 100)
coef <- rnorm(n = 10)
result <- spline(x, coef)

# evaluate transformation density
pdf_evaluations <- dtramo(result$fx, result$fx_deriv)

# evaluate transformation CDF
cdf_evaluations <- ptramo(result$fx)
}
