% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{equidistant_knots}
\alias{equidistant_knots}
\title{Create equidistant knots.}
\usage{
equidistant_knots(x, p = 10, l = 3, extend_range_factor = 0.01)
}
\arguments{
\item{x}{Input vector}

\item{p}{Desired number of parameters.}

\item{l}{Order of the spline. Defaults to \code{l=3} for cubic splines.}

\item{extend_range_factor}{Slightly extends the range over \code{min(x)} and \code{max(x)},
by setting \code{xlow <- min(x) - (max(x) - min(x)) * extend_range_factor}}
}
\value{
Vector of knots. Includes the outer knots. Will be of length \code{p + l + 1}.
}
\description{
Create equidistant knots.
}
