% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{get.seq}
\alias{get.seq}
\title{Import a Protein Sequence from a Database}
\usage{
get.seq(id, db = 'uniprot', as.string = TRUE)
}
\arguments{
\item{id}{the identifier of the protein of interest.}

\item{db}{a character string specifying the desired database; it must be one of 'uniprot' or 'metosite'.}

\item{as.string}{logical, if TRUE the imported sequence will be returned as a character string.}
}
\value{
Returns a protein  sequence either as a character vector or a as a character string.
}
\description{
Imports a protein sequence from a selected database.
}
\details{
MetOSite uses the same type of protein ID than UniProt.
}
\examples{
get.seq('P01009')
}
