% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go.enrich}
\alias{go.enrich}
\title{GO Terms Enrichment Tests}
\usage{
go.enrich(s_file, bg_file, aspect = 'BP', n = 20)
}
\arguments{
\item{s_file}{path to the file containing the list of identifiers for the sample of interest.}

\item{bg_file}{path to the file containing the list of IDs acting as background.}

\item{aspect}{character string indicating the aspect or sub-ontology. It must be one of 'BP' (Biological Process), 'MF' (Molecular Function) or 'CC' (Cellular Component). acting as background.}
}
\value{
Returns the results of the enrichement test as a dataframe.
}
\description{
Carry out GO term enrichment tests
}
\details{
It is essential that the items in the 'sample' vector correspond to items within the background.
}
\examples{
\dontrun{go.enrich("../bench/sample.txt", "../bench/background.txt", 'CC', n = 10)}
}
\references{

}
\seealso{
search.go(), term.go(), get.go(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
