% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{gorilla}
\alias{gorilla}
\title{GO Enrichment Analysis}
\usage{
gorilla(target, background = NULL, mode = 'mhg', db = 'proc', pvalue = 0.001, species = 'Homo sapiens')
}
\value{

}
\description{
Performs GO terms enrichment analyses.
}
\references{
Eden et al. (2009) BMC Bioinformatics 10:48.

Rhee et al. (2008) Nature Reviews Genetics 9:509–515.
}
\seealso{

}
\author{

}
