% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gracefully.R
\name{gracefully_fail}
\alias{gracefully_fail}
\title{Check that Internet Resource Work Properly and Fail Gracefully When Not}
\usage{
gracefully_fail(call, timeout = 10,  ...)
}
\arguments{
\item{call}{url of the resource.}

\item{timeout}{set maximum request time in seconds.}

\item{...}{further named parameters, such as query, headers, etc.}
}
\value{
The response object or NULL when the server does not respond properly.
}
\description{
Checks that internet resource works properly and fail gracefully when not.
}
\details{
To be used as an ancillary function.
}
\examples{
gracefully_fail("http://httpbin.org/delay/2")
}
\references{
https://gist.github.com/thefactmachine/18279b7796c0836d9188
}
\author{
thefactmachine
}
