% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{hdfisher.go}
\alias{hdfisher.go}
\title{Hypothesis-driven Fisher Test}
\usage{
hdfisher.go(target, background, query)
}
\arguments{
\item{target}{either a vector containing the UniProt IDs of the target set or the path to the txt file containing the list of IDs.}

\item{background}{a dataframe with two columns (Uniprot ID and GO terms) and as many rows as different proteins there are in the background set.}

\item{query}{character string defining the query.}

\item{analysis}{a character string indicating whether the desired analysis is the enrichment ('enrichment') or depletion ('depletion').}
}
\value{
Returns...
}
\description{
Carries out an enrichment Fisher's test using a hypothesis driven approach.
}
\examples{
hdfisher.go(c('Q14667', 'Q5JSZ5'), background.go(c("Q13015", "Q14667", "P08575", "Q5JSZ5", "P13196")), 'extracellular')
}
\references{
Rhee et al. (2008) Nature Reviews Genetics 9:509–515.
}
\seealso{
search.go(), term.go(), get.go(), background.go(), go.enrich(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
