% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{id.mapping}
\alias{id.mapping}
\title{Identifier Mapping}
\usage{
id.mapping(id, from, to)
}
\arguments{
\item{id}{the identifier to be converted.}

\item{from}{the type for the identifier of origin; it must be one of 'uniprot', 'pdb', or 'kegg'.}

\item{to}{the type for the identifier of destination; it must be one of 'uniprot', 'pdb', or 'kegg'.}
}
\value{
Returns a character string corresponding to the requested identifier.
}
\description{
Mapping between protein identifiers.
}
\examples{
\dontrun{id.mapping('P01009', from = 'uniprot', to = 'pdb')}
}
\seealso{
pdb2uniprot(), uniprot2pdb()
}
\author{
Juan Carlos Aledo
}
