% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{kegg.uniprot}
\alias{kegg.uniprot}
\title{Identifier Mapping From KEGG to UniProt}
\usage{
kegg.uniprot(id)
}
\arguments{
\item{id}{the identifier to be converted.}
}
\value{
Returns a character string corresponding to the requested identifier.
}
\description{
Mapping between KEGG and UniProt protein identifiers.
}
\examples{
\dontrun{kegg.uniprot('hsa:5265')}
}
\seealso{
id.mapping()
}
\author{
Juan Carlos Aledo
}
