% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{net.go}
\alias{net.go}
\title{Gene Ontology Network}
\usage{
net.go(path2data, threshold = 0.2, silent = FALSE)
}
\arguments{
\item{threshold}{threshold value of the Jaccard index above which, two proteins are considered to be linked.}

\item{silent}{logical, if FALSE print details of the running process.}

\item{path2data}{path to the dataset file.}
}
\value{
a list containing (i) the dataframe corresponding to the computed Jaccard matrix and (ii) the adjacency matrix.
}
\description{
Explores the relationship among proteins from a given set.
}
\details{
Finds the adjacency matrix for the relation among proteins from a given set based on their GO terms. The input data can be either a txt file containing UniProt IDs for the set protein of interest (one ID per line) or, alternatively, a Rda file containing a Jaccard matrix previously computed.
}
\examples{
\dontrun{net.go(path2data = "./GOvivo.txt")}
}
\seealso{
search.go(), term.go(), get.go(), go.enrich(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
