% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onion_spline.R
\name{onion_coef}
\alias{onion_coef}
\title{Computes the coefficient increments of an onion spline.}
\usage{
onion_coef(log_increments, knots, shift = NULL)
}
\arguments{
\item{log_increments}{Unconstrained onion spline log increments.}

\item{knots}{A list defining the knots for the onion spline, most easily
created with \code{\link[=onion_knots]{onion_knots()}}.}

\item{shift}{Numeric scalar for normalizing the onion spline's slope. Most
easily created with \code{\link[=onion_shift]{onion_shift()}}. If \code{NULL} (default), the shift is
computed automatically to normalize the spline's average slope to one.}
}
\value{
Numeric vector of coefficient increments.
}
\description{
Note that these are the increments of the spline coefficients, not the
full coefficients. To evaluate the onion spline, you have two options:
1)
}
\examples{
knots <- onion_knots(-4, 4, p = 10)
log_inc <- rnorm(10)
coef <- onion_coef(log_inc, knots)

x <- runif(100, -4, 4)
B <- splines::splineDesign(knots$knots, x)

# Option 1: Compute cumulative sum of coefficients.
fx <- B \%*\% cumsum(coef)

# Option 2: Absorb cumulative sum into design matrix.

}
