% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onion_spline.R
\name{onion_knots}
\alias{onion_knots}
\title{Create equidistant knots for an onion spline of order 3.}
\usage{
onion_knots(a_inner, b_inner, p, extend_range_factor = 0)
}
\arguments{
\item{a_inner}{Inner left boundary}

\item{b_inner}{Inner right boundary}

\item{p}{Desired number of parameters.}

\item{extend_range_factor}{Slightly extends the range over \code{min(x)} and \code{max(x)},
by setting \code{xlow <- min(x) - (max(x) - min(x)) * extend_range_factor}}
}
\value{
List of attributes.
}
\description{
Create equidistant knots for an onion spline of order 3.
}
