% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onion_spline.R
\name{onion_shift}
\alias{onion_shift}
\title{Computes the constant shift for onion spline coefficients.}
\usage{
onion_shift(log_increments, knots)
}
\arguments{
\item{log_increments}{Unconstrained onion spline log increments.}

\item{knots}{A list defining the knots for the onion spline, most easily
created with \code{\link[=onion_knots]{onion_knots()}}.}
}
\value{
A numeric scalar, the onion shift.
}
\description{
Subtracting this shift from unconstrained log increments ensures an average
slope of one for the onion spline.
}
\examples{
knots <- onion_knots(-4, 4, p = 10)
log_inc <- rnorm(10)
shift <- onion_shift(log_inc, knots)
}
