% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{pairwise.dist}
\alias{pairwise.dist}
\title{Compute Euclidean Distances}
\usage{
pairwise.dist(a, b, squared = TRUE)
}
\arguments{
\item{a, b}{matrices (NxD) and (MxD), respectively, where each row represents a D-dimensional point.}

\item{squared}{return containing squared Euclidean distance}
}
\value{
Euclidean distance matrix (NxM). An attribute "squared" set to the
value of param \code{squared} is provided.
}
\description{
Computes the pairwise distance matrix between two sets of points
}
\examples{
pairwise.dist(matrix(1:9, ncol = 3), matrix(9:1, ncol = 3))
}
