% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_ancillary.R
\name{pdb.chain}
\alias{pdb.chain}
\title{Download and/or Split PDB Files.}
\usage{
pdb.chain(pdb, keepfiles = FALSE)
}
\arguments{
\item{pdb}{the path to the PDB of interest or a 4-letter identifier.}

\item{keepfiles}{logical, if TRUE the function makes a 'temp' directory within the current directory and save in it a pdb file for each chain present in the given structure.}
}
\value{
The function returns a chr vector where each coordinate is a chain from the structure.
}
\description{
Downloads a PDB file (if required) and splits it to provide a file by chain.
}
\examples{
pdb.chain('1bpl')
}
\author{
Juan Carlos Aledo
}
