% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{pdb.quaternary}
\alias{pdb.quaternary}
\title{Protein Subunit Composition}
\usage{
pdb.quaternary(pdb, keepfiles = FALSE)
}
\arguments{
\item{pdb}{the path to the PDB of interest or a 4-letter identifier.}

\item{keepfiles}{logical, if TRUE the fasta file containing the alignment of the subunits is saved in the current directory, as well as the split pdb files.}
}
\value{
This function returns a list with four elements: (i) a distances matrix, (ii) the sequences, (iii) chains id, (iv) the PDB ID used.
}
\description{
Determines the subunit composition of a given protein.
}
\details{
A fasta file containing the alignment among the subunit sequences can be saved in the current directory if required.
}
\examples{
\dontrun{pdb.quaternary('1bpl')}
}
\author{
Juan Carlos Aledo
}
