% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{pdb.seq}
\alias{pdb.seq}
\title{Get Chain Sequences}
\usage{
pdb.seq(pdb)
}
\arguments{
\item{pdb}{the 4-letter PDB identifier.}
}
\value{
Returns a dataframe with as many rows as different chains are present in the PDB. For each row six variables are returned: (i) the entry id, (ii) the entity id, (iii) the chain, (iv) the protein name, (v) the species and (vi) the sequence.
}
\description{
Gets the sequences of the chain find in a given PDB.
}
\examples{
\dontrun{pdb.seq('1bpl')}
}
\author{
Juan Carlos Aledo
}
