% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmScan.R
\name{reg.scan}
\alias{reg.scan}
\title{Scan a Protein in Search of Regulatory PTM Sites}
\usage{
reg.scan(up_id)
}
\arguments{
\item{up_id}{a character string corresponding to the UniProt ID.}
}
\value{
Returns a dataframe where each row corresponds to a residue, and the columns inform about the regulatory modifications.
}
\description{
Scans the indicated protein in search of regulatory PTM sites.
}
\examples{
\dontrun{reg.scan('P01009')}
}
\references{
Hornbeck et al. Nucleic Acids Res. 2019 47:D433-D441, (PMID: 30445427).
}
\seealso{
meto.scan(), ac.scan(), me.scan(), ub.scan(), su.scan(), gl.scan(), sni.scan(), ni.scan(), ptm.scan(), p.scan(), dis.scan()
}
\author{
Juan Carlos Aledo
}
