% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.R
\name{renum.pdb}
\alias{renum.pdb}
\title{Renumerate Residue Position}
\usage{
renum.pdb(pdb, chain, uniprot)
}
\arguments{
\item{pdb}{the PDB ID or the path to a pdb file.}

\item{chain}{the chain of interest.}

\item{uniprot}{the UniProt ID.}
}
\value{
Returns a dataframe containing the re-numerated sequence.
}
\description{
Renumerates residue position of a PDB sequence to match the corresponding UniProt sequence.
}
\examples{
renum.pdb(pdb = '121P', chain = 'A', uniprot = 'P01112')
}
\seealso{
is.at(), aa.at(), renum.meto(), renum(), aa.compo()
}
\author{
Juan Carlos Aledo
}
