% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{search.go}
\alias{search.go}
\title{Search a Simple User Query}
\usage{
search.go(query)
}
\arguments{
\item{query}{character string defining the query.}
}
\value{
Returns a ......
}
\description{
Searchs a simple user query.
}
\details{
that....
}
\examples{
search.go('oxidative stress')
}
\seealso{
term.go(), get.go(), go.enrich(), gorilla(), net.go()
}
\author{
Juan Carlos Aledo
}
