% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmScan.R
\name{sni.scan}
\alias{sni.scan}
\title{Scan a Protein in Search of S-nitrosylation Sites}
\usage{
sni.scan(up_id, db = 'all')
}
\arguments{
\item{up_id}{a character string corresponding to the UniProt ID.}

\item{db}{the database where to search. It should be one among 'PSP', 'dbPTM', 'all'.}
}
\value{
Returns a dataframe where each row corresponds to a modifiable residue.
}
\description{
Scans the indicated protein in search of S-nitrosylation sites.
}
\examples{
\dontrun{sni.scan('P01009')}
}
\references{
Huang et al. Nucleic Acids Res. 2019 47:D298-D308, (PMID: 30418626).
}
\seealso{
meto.scan(), ac.scan(), me.scan(), ub.scan(), su.scan(), gl.scan(), p.scan(), ni.scan(), ptm.scan(), reg.scan(), dis.scan()
}
\author{
Juan Carlos Aledo
}
