% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{species.mapping}
\alias{species.mapping}
\title{Map Protein ID to Species}
\usage{
species.mapping(id, db = 'uniprot')
}
\arguments{
\item{id}{the identifier of the protein of interest.}

\item{db}{a character string specifying the corresponding database. Currently, only 'uniprot' or 'pdb' are valid options.}
}
\value{
Returns a character string identifying the organism to which the given protein belong.
}
\description{
Maps a protein ID to its corresponding organism.
}
\examples{
species.mapping('P01009')
}
\seealso{
id.features()
}
\author{
Juan Carlos Aledo
}
