% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getseq.R
\name{uniprot.kegg}
\alias{uniprot.kegg}
\title{Identifier Mapping From UniProt to KEGG}
\usage{
uniprot.kegg(id)
}
\arguments{
\item{id}{the identifier to be converted.}
}
\value{
Returns a character string corresponding to the requested identifier.
}
\description{
Mapping between UniProt and KEGG protein identifiers.
}
\examples{
\dontrun{uniprot.kegg('P01009')}
}
\seealso{
id.mapping()
}
\author{
Juan Carlos Aledo
}
