% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsplines.R
\name{get_extrap_bspline_and_deriv}
\alias{get_extrap_bspline_and_deriv}
\title{Generates a function to evaluate a B-spline and its derivative with
extrapolation.}
\usage{
get_extrap_bspline_and_deriv(
  basis_and_deriv_fn,
  target_left = c("identity"),
  target_right = c("identity")
)
}
\arguments{
\item{basis_and_deriv_fn}{A function that follows the signature of the
function returned by [get_approx_basis_and_deriv_fun()].}

\item{target_left, target_right}{Strings, indicating how to extrapolate for
`x` outside the range supported by the spline's knots. Supported values:
* `"identity"`: Extrapolates with the identity function.}
}
\value{
A function that takes to arguments: `x`, a numeric vector of values
     at which to evaluate the spline, and `coef`, a numeric vector of spline
     coefficients. `length(coef)` must equal the number of columns of the
     basis matrix returned by `basis_and_deriv_fn`. The function returns a
     list of two numeric vectors:
     * `fx`: Numeric vector of spline evaluations.
     * `fx_deriv`: Numeric vector of spline derivative evaluations.
}
\description{
Generates a function to evaluate a B-spline and its derivative with
extrapolation.
}
\examples{
knots <- equidistant_knots(c(-4, 4), p = 10)
approx_basis_and_deriv <- get_approx_basis_and_deriv_fun(knots, ngrid = 1000)
spline <- get_extrap_bspline_and_deriv(approx_basis_and_deriv)

# evaluate spline
x <- rnorm(n = 100)
coef <- rnorm(n = 10)
result <- spline(x, coef)
}
